anti_gravity_system/requirements.txt
pydantic
openai
chromadb
pyyaml
rich
fastapi
uvicorn
